MDL-63791 atto_recordrtc: Convert upload file size limit to bytes
When the upload file size limit is being set via the
upload_max_filesize directive, the plugin incorrectly parses
the value: 2G as 2 MB.

This is caused by the plugin expecting this value in the megabytes
unit.

The fix here converts the value to bytes and removes the megabyte to
bytes conversion.